refactoring, add readmy

+++ user-dashboard/components/data-table/DataTable.tsx
@@ -24,6 +24,7 @@ import { Button } from "../ui/button";
 import { toast } from "sonner"
 import UploadFile from "../ui/uploadBtn";
 import WaitingDialog from "../WaitingDialog";
+import ConfirmDialog from "../ConfirmDialog";
 
 export function DataTable<TData, TValue>() {
   const [sorting, setSorting] = React.useState<SortingState>([])
@@ -32,22 +33,25 @@ export function DataTable<TData, TValue>() {
   const [isWaiting, setIsWaiting] = useState(false);
   const [isDialogOpen, setIsDialogOpen] = React.useState<boolean>(false);
   const [selectedUser, setSelectedUser] = React.useState<User | null>(null);
-
+  const [isConfirm, setIsConfirm] = useState(false);
 
   const queryClient = useQueryClient();
 
-  const { data, isLoading } = useQuery<UserTableData>({
+  const { data, isLoading, isError } = useQuery<UserTableData>({
     queryKey: ["users", page, sorting],
     queryFn: () => getUsers(page, sorting),
   });
 
+  if (isError) {
+    return (<div>Помилка отримання данних. Спробуйте пізніше.</div>);
+  }
   
 
   const onDeleteSuccess = async (user: User) => {
-    console.log(`Користувач ${user.user_name} був видаленний`);
     queryClient.invalidateQueries({ queryKey: ["users", page, sorting] });
     onOpenChange(false);
-    setIsWaiting(false)
+    setIsWaiting(false);
+    toast.success("Користувача успішно видалено.");
   };
   const onRequestError = () => {
     toast.error("Ой! Щось пішло не так", {
@@ -65,30 +69,27 @@ export function DataTable<TData, TValue>() {
   const onDelete = useCallback((user: User) => {
     setIsWaiting(true)
     delMut.mutate(user.id)
-  }, []);
+  }, [queryClient, page, sorting, delMut]);
 
   const onEdit = useCallback((user: User) => {
     setSelectedUser(user);
     setIsDialogOpen(true);
+    toast.success("Користувача успішно відредаговано.");
 
-  }, []);
+  }, [isDialogOpen, selectedUser]);
 
   const onUploadEnd = useCallback((value: boolean) => {
     queryClient.invalidateQueries({ queryKey: ["users", page, sorting] });
-  }, []);
+  }, [queryClient, page, sorting]);
 
   const onOpenChange = useCallback((value: boolean) => {
     setIsDialogOpen(value);
 
     if (!value) {
       setSelectedUser(null);
-      console.log("selectedUser", selectedUser)
       queryClient.invalidateQueries({ queryKey: ["users", page, sorting] });
-
-    } else {
-      console.log("selectedUser", selectedUser)
     }
-  }, [])
+  }, [queryClient, page, sorting])
   const columns = useMemo(() => UserColumns({ onEdit, onDelete }), []);
 
   const totalPages = data?.total ? Math.ceil(data.total / 10) : 0;
@@ -107,33 +108,41 @@ export function DataTable<TData, TValue>() {
     rowCount: data?.total ? data.total : 0,
   })
 
-  const nextPage = () => {
-    setPage(page + 1)
-  }
-  const previousPage = () => {
-    setPage(page - 1)
-  }
+  const changePage = (newPage: number) => {
+    if (newPage >= 0 && newPage < totalPages) {
+      setPage(newPage);
+    }
+  };
 
   const onDeleteAll = async (event: any) => {
-    setIsDeletingAll(true)
-    setIsWaiting(true)
-    await deleteAllUsers()
-    queryClient.invalidateQueries({ queryKey: ["users", page, sorting] });
-    setIsDeletingAll(false)
-    setIsWaiting(false)
-    toast("Видалив усе!", {
-      description: "Sunday, December 03, 2023 at 9:00 AM",
-      action: {
-        label: "Undo",
-        onClick: () => console.log("Undo"),
-      },
-    })
+    event.preventDefault();
+    setIsConfirm(true);
+    
+    
   }
+  const closeEvent = async (val: boolean) => {
+    if(val) {
+      try {
+        setIsDeletingAll(true);
+        setIsWaiting(true);
+        await deleteAllUsers();
+        queryClient.invalidateQueries({ queryKey: ["users", page, sorting] });
+      } catch (error) {
+        toast.error("Ошибка при удалении всех пользователей.");
+      } finally {
+        setIsDeletingAll(false);
+        setIsWaiting(false);
+        toast.success("Усі користувачі успішно видалені.");
+      }
+    }
+    setIsConfirm(false);
+  };
 
   return (
 
     <div className="conteiner">
       <WaitingDialog isOpen={isWaiting}/>
+      <ConfirmDialog isOpen={isConfirm} onCloseEvent={closeEvent} />
       <div className="flex flex-wrap items-center justify-center md:justify-start gap-4 py-5">
         <div className="flex items-center gap-4">
 
@@ -215,32 +224,32 @@ export function DataTable<TData, TValue>() {
           <Button
             variant="outline"
             size="sm"
-            onClick={() => setPage(0)}
-            disabled={page === 0}
+            onClick={() => changePage(0)}
+            disabled={page === 0 || totalPages === 0}
           >
             Перша сторінка
           </Button>
           <Button
             variant="outline"
             size="sm"
-            onClick={() => previousPage()}
-            disabled={page === 0}
+            onClick={() => changePage(page - 1)}
+            disabled={page === 0 || totalPages === 0}
           >
             Попередня
           </Button>
           <Button
             variant="outline"
             size="sm"
-            onClick={() => nextPage()}
-            disabled={(page + 1) >= totalPages}
+            onClick={() => changePage(page + 1)}
+            disabled={(page + 1) >= totalPages || totalPages === 0}
           >
             Наступна
           </Button>
           <Button
             variant="outline"
             size="sm"
-            onClick={() => setPage(totalPages - 1)}
-            disabled={(page + 1) >= totalPages}
+            onClick={() => changePage(totalPages - 1)}
+            disabled={(page + 1) >= totalPages || totalPages === 0}
           >
             Остання сторінка
           </Button>
